- added .sonarcloud.properties, with configured exclusions - moved App.tsx and related files to ./src - updated android_build.yml to take google-services.json from github secrets - added lint coverage report in json format
