+++ .github/workflows/android_build.yml
@@ -47,6 +47,11 @@ jobs:
           echo "ANDROID_RELEASE_KEY_ALIAS=${{ secrets.ANDROID_ALIAS }}" >> $GITHUB_ENV
           echo "ANDROID_RELEASE_KEY_PASSWORD=${{ secrets.ANDROID_KEY_PASSWORD }}" >> $GITHUB_ENV
 
+      - name: Create Google Services JSON File
+        env:
+          GOOGLE_SERVICES_JSON: ${{ secrets.GOOGLE_SERVICES_JSON_BASE64 }}
+        run: echo $GOOGLE_SERVICES_JSON | base64 -di > ./android/app/google-services.json
+
       - uses: actions/cache@v3
         with:
           path: |
