+++ .sonarcloud.properties
@@ -1,0 +1,18 @@
+sonar.links.homepage=https://github.com/enhancer13/navi-home-client
+sonar.links.issue=https://github.com/enhancer13/navi-home-client/issues
+
+sonar.typescript.tsconfigPath=tsconfig.json
+sonar.sources=src
+sonar.tests=__tests__
+
+sonar.test.inclusions=**/__tests__/*.ts
+
+# code duplication
+sonar.cpd.exclusions=**/__tests__/**,**/PaperTheme.ts
+
+# coverage
+sonar.coverage.exclusions=**/__tests__/**
+sonar.typescript.lcov.reportPaths=coverage/lcov.info
+
+# external reporters integration
+sonar.eslint.reportPaths=eslint-results.json
