+++ package.json
@@ -23,8 +23,7 @@
     "lint:translations": "eslint ./src/translations/ --fix --ext .json  ",
     "test": "jest",
     "test-watch": "yarn run test --watch",
-    "test-ci": "yarn run test --coverage",
-    "test-badges": "yarn test:ci  && jest-coverage-badges --input coverage/coverage-summary.json --output __badges__"
+    "test-ci": "yarn run test --coverage"
   },
   "dependencies": {
     "@babel/runtime": "^7.22.5",
@@ -116,6 +115,7 @@
   },
   "license": "MIT",
   "jest": {
+    "coverageReporters": ["json-summary"],
     "preset": "react-native",
     "moduleFileExtensions": [
       "ts",
