+++ src/Components/Controls/ListItems/ListIcon.tsx
@@ -1,7 +1,7 @@
 import {List, useTheme} from "react-native-paper";
 import React from "react";
 import {StyleProp, StyleSheet, ViewStyle} from "react-native";
-import {NaviTheme} from "../../../../PaperTheme";
+import {NaviTheme} from "../../../PaperTheme";
 
 export type ListIconProps = {
     icon: string;
