+++ src/Features/ErrorBoundary/AppErrorBoundary.tsx
@@ -5,7 +5,7 @@ import { AuthenticationFailed } from "../../Errors/AuthenticationFailed";
 import { Props as FallbackComponentProps } from "react-native-error-boundary/src/ErrorBoundary/FallbackComponent";
 import { useNavigation } from "@react-navigation/native";
 import FallbackComponent from "react-native-error-boundary/lib/ErrorBoundary/FallbackComponent";
-import {RootNavigationProp} from "../../../RootStackNavigator";
+import {RootNavigationProp} from "../../RootStackNavigator";
 
 type AuthErrorBoundaryProps = {
     children: Exclude<NonNullable<ReactNode>, string | number | boolean>;
