+++ src/Screens/Home/MyAccount/AccountDetailsScreen.tsx
@@ -13,7 +13,7 @@ import {NativeScrollEvent} from "react-native/Libraries/Components/ScrollView/Sc
 import {AccountStackParamList} from "./index";
 import {heightPercentageToDP as hp, widthPercentageToDP as wp} from "react-native-responsive-screen";
 import {useApplicationSettings} from "../../../Features/DataStorage/Hooks/useApplicationSettings";
-import {NaviTheme} from "../../../../PaperTheme";
+import {NaviTheme} from "../../../PaperTheme";
 
 const PAGE_TITLE = 'My Account';
 const HEADER_HEIGHT = hp(5);
